Change N to int

+++ fft1d.C
@@ -7,7 +7,7 @@
 
 /*readonly*/ CProxy_Main mainProxy;
 /*readonly*/ int numChares;
-/*readonly*/ uint64_t N;
+/*readonly*/ int N;
 
 struct fftMsg : public CMessage_fftMsg {
   int source;
@@ -20,7 +20,7 @@ struct Main : public CBase_Main {
 
   Main(CkArgMsg* m) {
     numChares = atoi(m->argv[1]);
-    N = atol(m->argv[2]);
+    N = atoi(m->argv[2]);
     delete m;
 
     mainProxy = thisProxy;
@@ -41,7 +41,7 @@ struct Main : public CBase_Main {
   void FFTDone() {
     double time = CkWallTimer() - start;
     double gflops = 5 * (double)N*N * log2((double)N*N) / (time * 1000000000);
-    CkPrintf("chares: %d\ncores: %d\nsize: %ld\ntime: %f sec\nrate: %f GFlop/s\n",
+    CkPrintf("chares: %d\ncores: %d\nsize: %d\ntime: %f sec\nrate: %f GFlop/s\n",
              numChares, CkNumPes(), N*N, time, gflops);
 
     fftProxy.initValidation();
